Change the behavior of the error handler in requests
Adding a message output for the user in case of incorrect data input for the request.
Displaying a message about the lack of data in the body of the GET request and in case of connection problems with the server.
Implementation of requests by pressing the Enter key for ease of use.